+++ src/main/java/com/saintdan/framework/controller/ClientController.java
@@ -12,15 +12,22 @@ import com.saintdan.framework.enums.ErrorType;
 import com.saintdan.framework.enums.OperationStatus;
 import com.saintdan.framework.exception.CommonsException;
 import com.saintdan.framework.param.ClientParam;
+import com.saintdan.framework.po.Client;
 import com.saintdan.framework.po.User;
 import com.saintdan.framework.tools.QueryHelper;
-import com.saintdan.framework.vo.PageVO;
+import com.saintdan.framework.vo.ClientVO;
 import com.saintdan.framework.vo.ResultVO;
 import javax.validation.Valid;
+import net.kaczmarzyk.spring.data.jpa.domain.DateBetween;
+import net.kaczmarzyk.spring.data.jpa.domain.In;
+import net.kaczmarzyk.spring.data.jpa.domain.Like;
+import net.kaczmarzyk.spring.data.jpa.web.annotation.And;
+import net.kaczmarzyk.spring.data.jpa.web.annotation.Spec;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -49,10 +56,10 @@ public class ClientController {
    * @return {@link com.saintdan.framework.vo.ClientVO}
    */
   @RequestMapping(method = RequestMethod.POST)
-  public ResultVO create(@CurrentUser User currentUser, ClientParam param) {
+  public ResultVO create(@CurrentUser User currentUser, @Valid ClientParam param, BindingResult result) {
     try {
       // Validate current user, param and sign.
-      ResultVO resultVO = validateHelper.validate(currentUser, param, logger);
+      ResultVO resultVO = validateHelper.validateWithOutSignCheck(result, currentUser, logger);
       if (resultVO != null) {
         return resultVO;
       }
@@ -68,23 +75,23 @@ public class ClientController {
   }
 
   /**
-   * Show {@link com.saintdan.framework.vo.ClientVO} in {@link com.saintdan.framework.vo.PageVO}.
+   * Show all.
    *
    * @param param {@link ClientParam}
-   * @return {@link com.saintdan.framework.vo.ClientVO} in
+   * @return all clients.
    */
   @RequestMapping(method = RequestMethod.GET)
-  public ResultVO show(ClientParam param) {
+  public ResultVO all(
+      @And({
+          @Spec(path = "name", spec = Like.class),
+          @Spec(path = "validFlag", constVal = "VALID", spec = In.class),
+          @Spec(path = "createdDate", params = {"createdDateAfter, createdDateBefore"}, spec = DateBetween.class)}) Specification<Client> clientSpecification,
+      ClientParam param) {
     try {
-      // Sign validate.
-      ResultVO resultVO = validateHelper.validate(param, logger);
-      if (resultVO != null) {
-        return resultVO;
-      }
       if (param.getPageNo() == null) {
-        return resultHelper.successResp(clientDomain.getAll());
+        return resultHelper.successResp(clientDomain.getAll(clientSpecification, QueryHelper.getSort(param.getSortBy()), ClientVO.class));
       }
-      return resultHelper.successResp(clientDomain.getPage(QueryHelper.getPageRequest(param), PageVO.class));
+      return resultHelper.successResp(clientDomain.getPage(clientSpecification, QueryHelper.getPageRequest(param), ClientVO.class));
     } catch (CommonsException e) {
       // Return error information and log the exception.
       return resultHelper.infoResp(logger, e.getErrorType(), e.getMessage());
@@ -101,19 +108,12 @@ public class ClientController {
    * @return {@link com.saintdan.framework.vo.ClientVO}
    */
   @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-  public ResultVO show(@PathVariable String id, String sign) {
+  public ResultVO detail(@PathVariable String id) {
     try {
       if (StringUtils.isBlank(id)) {
         return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, ControllerConstant.ID_PARAM));
       }
-      ClientParam param = new ClientParam(Long.valueOf(id));
-      param.setSign(sign);
-      // Sign validate.
-      ResultVO resultVO = validateHelper.validate(param, logger);
-      if (resultVO != null) {
-        return resultVO;
-      }
-      return resultHelper.successResp(clientDomain.getById(ResultVO.class, param));
+      return resultHelper.successResp(clientDomain.getById(new ClientParam(Long.valueOf(id)), ResultVO.class));
     } catch (CommonsException e) {
       // Return error information and log the exception.
       return resultHelper.infoResp(logger, e.getErrorType(), e.getMessage());
@@ -137,7 +137,7 @@ public class ClientController {
         return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, ControllerConstant.ID_PARAM));
       }
       // Validate current user, param and sign.
-      ResultVO resultVO = validateHelper.validate(result, currentUser, param, logger);
+      ResultVO resultVO = validateHelper.validateWithOutSignCheck(result, currentUser, logger);
       if (resultVO != null) {
         return resultVO;
       }
@@ -159,15 +159,13 @@ public class ClientController {
    * @return {@link com.saintdan.framework.vo.ClientVO}
    */
   @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-  public ResultVO delete(@CurrentUser User currentUser, @PathVariable String id, String sign) {
+  public ResultVO delete(@CurrentUser User currentUser, @PathVariable String id) {
     try {
       if (StringUtils.isBlank(id)) {
         return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, ControllerConstant.ID_PARAM));
       }
       ClientParam param = new ClientParam(Long.valueOf(id));
-      param.setSign(sign);
-      // Sign validate.
-      ResultVO resultVO = validateHelper.validate(param, logger);
+      ResultVO resultVO = validateHelper.validateWithOutSignCheck(currentUser, logger);
       if (resultVO != null) {
         return resultVO;
       }